Prevent ADC rollover

+++ source/Core/BSP/Pinecilv2/bl_mcu_sdk/drivers/bl702_driver/std_drv/inc/bl702_adc.h
@@ -271,7 +271,7 @@ typedef struct {
 typedef struct {
   int8_t   posChan; /*!< Positive channel */
   int8_t   negChan; /*!< Negative channel */
-  uint16_t value;   /*!< ADC value */
+  uint32_t value;   /*!< ADC value */
   // float    volt;    /*!< ADC voltage result */
 } ADC_Result_Type;
